Update to use systemd service units
Still requires mountpoints to be specified in /etc/fstab, only they
should be commented out. The script will then automatically create
systemd service units for the mount points that the automountnfs then
controls. Systemd auto mounting is nice, but finicky, and when it
doesn't work it slows down boot which is annoying.